Mollified MSVC's stupidity.

+++ Working/Source/Graphics/Environment/Window.cpp
@@ -13,7 +13,7 @@ Window::Window(uint32_t width, uint32_t height, const char* name) {
     glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
     glfwWindowHint(GLFW_RESIZABLE, GLFW_FALSE);
 
-    window = glfwCreateWindow(width, height, name, nullptr, nullptr);
+    window = glfwCreateWindow((int)width, (int)height, name, nullptr, nullptr);
 }
 
 /**
